+++ src/storage/index.ts
@@ -1,3 +1,4 @@
 export * from "./chromeStorage";
 export * from "./keys";
+export * from "./localMirror";
 export * from "./repositories";